Move cpp interface to subfolder
