+++ interfaces/acados_cpp/ocp_qp/hpipm_helper.hpp
