+++ swig/ocp_qp.i
@@ -7,9 +7,9 @@
 #include <vector>
 
 #include "acados_c/ocp_qp_interface.h"
-#include "acados_cpp/ocp_qp.hpp"
-#include "acados_cpp/ocp_qp_solution.hpp"
-#include "acados_cpp/options.hpp"
+#include "acados_cpp/ocp_qp/ocp_qp.hpp"
+#include "acados_cpp/ocp_qp/ocp_qp_solution.hpp"
+#include "acados_cpp/ocp_qp/options.hpp"
 
 #include "acados/ocp_qp/ocp_qp_common.h"
 #include "acados/utils/print.h"
@@ -96,13 +96,13 @@ LangObject *ocp_qp_output(const ocp_qp_in *in, const ocp_qp_out *out) {
 %}
 
 %ignore operator<<;
-%include "acados_cpp/options.hpp"
+%include "acados_cpp/ocp_qp/options.hpp"
 
 %rename("$ignore", %$isconstructor) ocp_qp_solution;
-%include "acados_cpp/ocp_qp_solution.hpp"
+%include "acados_cpp/ocp_qp/ocp_qp_solution.hpp"
 %ignore extract;
 %rename("$ignore", %$isconstructor) ocp_qp;
-%include "acados_cpp/ocp_qp.hpp"
+%include "acados_cpp/ocp_qp/ocp_qp.hpp"
 
 %rename("%s", %$isconstructor) ocp_qp;
 %rename("%s") extract;
